Use pg_ctl on windows
